requirements.txt
flask
flask-bcrypt
flask-api
flask-cors
flask-migrate
flask-script
psycopg2
PyJWT
coverage
flasgger
coveralls
flask-mail==0.9.0
